Revert build to pre-c++11 so CentOS 6.6 will compile without upgrading gcc.

+++ branches/flame_x_dev/include/api/SeaBreezeWrapper.h
@@ -97,7 +97,7 @@ public:
     int    getElectricDarkPixelIndices(int index, int *errorCode, int *indices, int length);
 	int    getOpticalDarkPixelIndices(int index, int *errorCode, int *indices, int length);
 	int    getActivePixelIndices(int index, int *errorCode, int *indices, int length);
-	int    getNumberOfPixels(int index, int *errorCode);
+	unsigned short    getNumberOfPixels(int index, int *errorCode);
     void   setTriggerMode            (int index, int *errorCode, int mode);
     void   setStrobeEnable           (int index, int *errorCode, unsigned char strobe_enable);
     void   setShutterOpen            (int index, int *errorCode, unsigned char opened);
@@ -618,7 +618,7 @@ extern "C" {
 	* Note that not all spectrometers provide the number of detector pixels; in that case,
 	* this function will return zero.
 	*/
-	DLL_DECL unsigned int seabreeze_get_number_of_pixels(int index, int *error_code);
+	DLL_DECL unsigned short seabreeze_get_number_of_pixels(int index, int *error_code);
 
     /**
      * @brief This fills in the provided array (up to the given length) with the indices
