+++ branches/flame_x_dev/include/vendors/OceanOptics/features/spectrometer/OOISpectrometerFeatureInterface.h
@@ -84,7 +84,7 @@ namespace seabreeze {
         virtual long getIntegrationTimeMaximum() const = 0;
         virtual long getIntegrationTimeIncrement() const = 0;
 
-        virtual int getNumberOfPixels() const = 0;
+        virtual unsigned short getNumberOfPixels() const = 0;
         virtual int getMaximumIntensity() const = 0;
 
     };
